Scraping Program added
You will need to input the API key from the NY Times API in the modifiable code, but it should successfully scrape text, and write articles from the NY Times to be used as documents in a corpus. In order to get all of the documents needed, dates will need to be modified; the begin date will change based on the last article that was pulled and the end date will remain constant.

The API will allow us to use different keywords as well. For now, we have set the program to search for articles with "Williamsburg" and "change" in the body of the text.